requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
tensorflow-cpu==2.16.1
numpy==1.24.3
pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
Jinja2==3.1.2
click==8.1.7
blinker==1.6.3
bcrypt==4.0.1